core: split up core.h, installment 1
