+++ core/Double.carp
@@ -1,3 +1,5 @@
+(system-include "double.h")
+
 (defmodule Double
   (def pi 3.141592653589793)
   (def e 2.718281828459045)
