+++ core/Float.carp
@@ -1,3 +1,5 @@
+(system-include "float.h")
+
 (defmodule Float
   (def pi 3.1415926536f)
   (register neg (Fn [Float] Float))
