+++ core/Int.carp
@@ -1,3 +1,5 @@
+(system-include "int.h")
+
 (defmodule Int
   (register + (λ [Int Int] Int))
   (register - (λ [Int Int] Int))
