+++ core/long.h
@@ -1,0 +1,70 @@
+long Long__PLUS_(long x, long y)   { return x + y; }
+long Long__MINUS_(long x, long y)  { return x - y; }
+long Long__MUL_(long x, long y)    { return x * y; }
+long Long__DIV_(long x, long y)    { return x / y; }
+bool Long_safe_MINUS_add(long x, long y, long* res) { return __builtin_saddl_overflow(x, y, res); }
+bool Long_safe_MINUS_sub(long x, long y, long* res) { return __builtin_ssubl_overflow(x, y, res); }
+bool Long_safe_MINUS_mul(long x, long y, long* res) { return __builtin_smull_overflow(x, y, res); }
+bool Long__EQ_(long x, long y)     { return x == y; }
+bool Long__LT_(long x, long y)    { return x < y; }
+bool Long__GT_(long x, long y)    { return x > y; }
+
+long Long_inc(long x) { return x + 1; }
+long Long_dec(long x) { return x - 1; }
+long Long_abs(long x) { return labs(x); }
+long Long_bit_MINUS_shift_MINUS_left(long x, long y) { return x << y; }
+long Long_bit_MINUS_shift_MINUS_right(long x, long y) { return x >> y; }
+long Long_bit_MINUS_and(long x, long y) { return x & y; }
+long Long_bit_MINUS_or(long x, long y) { return x | y; }
+long Long_bit_MINUS_xor(long x, long y) { return x ^ y; }
+long Long_bit_MINUS_not(long x) { return ~x; }
+
+int Long_copy(long *x) { return *x; }
+
+long Long_from_MINUS_string(string *s) {
+    return atol(*s);
+}
+
+long Long_mod(long x, long divider) {
+    return x % divider;
+}
+
+void Long_seed(long seed) {
+    srand(seed);
+}
+
+long Long_random() {
+    return rand();
+}
+
+long Long_random_MINUS_between(long lower, long upper) {
+    long diff = upper - lower;
+    return lower + (rand() % diff);
+}
+
+string Long_str(long x) {
+    int size = snprintf(NULL, 0, "%ldl", x)+1;
+    string buffer = CARP_MALLOC(size);
+    snprintf(buffer, size, "%ldl", x);
+    return buffer;
+}
+
+string Long_format(string* str, long x) {
+    int size = snprintf(NULL, 0, *str, x)+1;
+    string buffer = CARP_MALLOC(size);
+    snprintf(buffer, size, *str, x);
+    return buffer;
+}
+
+bool Long_mask(long a, long b) {
+    return a & b;
+}
+
+int Long_to_MINUS_int(long a) {
+  return (int) a;
+}
+
+long Long_from_MINUS_int(int a) {
+  return (long) a;
+}
+
